# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Telemedicina para Todos API",
        "description": "Documentação da API do projeto Telemedicina para Todos. Endpoints internos (sessão) e preparação para API pública de interoperabilidade.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Servidor da aplicação (base atual)"
        },
        {
            "url": "/api/v1",
            "description": "API pública versionada (futuro – interoperabilidade)"
        }
    ],
    "paths": {
...
        "/api/notifications": {
            "get": {
                "tags": [
                    "Notificações"
                ],
                "summary": "Listar notificações",
                "description": "Listar notificações do usuário",
                "operationId": "911565ad9aff9ba7eded3f189509e578",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "required": false
                    },
                    {
                        "name": "unread_only",
                        "in": "query",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista paginada"
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                }
            }
        },
        "/api/notifications/unread": {
            "get": {
                "tags": [
                    "Notificações"
                ],
                "summary": "Notificações não lidas",
                "description": "Obter notificações não lidas",
                "operationId": "bbdf4779601a9c7758ebf68b0aedeb83",
                "responses": {
                    "200": {
                        "description": "Lista e contagem"
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                }
            }
        },
        "/api/notifications/unread-count": {
            "get": {
                "tags": [
                    "Notificações"
                ],
                "summary": "Contagem de não lidas",
                "description": "Obter contador de notificações não lidas",
                "operationId": "0272623bed30dd4d72141741fb3f25da",
                "responses": {
                    "200": {
                        "description": "count"
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                }
            }
        },
        "/api/notifications/{id}/read": {
            "post": {
                "tags": [
                    "Notificações"
                ],
                "summary": "Marcar como lida",
                "description": "Marcar notificação como lida",
                "operationId": "ccc13a8ee59dfc08e9b1f80e075ebfc1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da notificação",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Notificação marcada como lida"
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "404": {
                        "description": "Não encontrada"
                    }
                }
            }
        },
        "/api/notifications/read-all": {
            "post": {
                "tags": [
                    "Notificações"
                ],
                "summary": "Marcar todas como lidas",
                "description": "Marcar todas as notificações como lidas",
                "operationId": "80760dfd47d7dd101c507f26ca9f0de9",
                "responses": {
                    "200": {
                        "description": "Quantidade marcada"
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                }
            }
        },
        "/api/notifications/{id}": {
            "get": {
                "tags": [
                    "Notificações"
                ],
                "summary": "Exibir notificação",
                "description": "Obter uma notificação específica",
                "operationId": "f2edf725f41dd6c9a9338a6a7b39d779",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID da notificação",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Notificação"
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "404": {
                        "description": "Não encontrada"
                    }
                }
            }
        },
...
    },
...
}